fix: emoji picker visible logic & allow hide when click outside
